requirements.txt
openai==1.54.4
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.33.0
